+++ app/src/main/java/com/sipos/kebabsk/feature/expense/data/remote/OperationalExpenseApiService.kt
@@ -8,7 +8,7 @@ import retrofit2.http.POST
 import retrofit2.http.Url
 
 data class OperationalExpenseRequest(
-    @SerializedName("amount") val amount: Double,
+    @SerializedName("amount") val amount: Long,
     @SerializedName("source") val source: String,
     @SerializedName("note") val note: String?
 )
